Fix a couple of compile warnings with gcc4.3.2.
For constant strings it is better to use fputs rather than fprintf. This 
prevents inadvertently or maliciously including format characters such 
as %d in the string which would cause fprintf to fail.

Also, on 64-bit platforms size_t may not be an int so explicitly cast 
the value to an int to prevent mismatch with %d format. Newer glibc 
versions include a z length modifier which signifies size_t, but unless 
this is widely available it is likely only to cause additional problems. 


svn path=/trunk/; revision=9110

+++ src/plbuf.c
@@ -924,7 +924,7 @@ wr_command(PLStream *pls, U_CHAR c)
         pls->plbuf_buffer_size += pls->plbuf_buffer_grow;
 
         if (pls->verbose)
-            printf("Growing buffer to %d KB\n", pls->plbuf_buffer_size / 1024);
+            printf("Growing buffer to %d KB\n", (int)(pls->plbuf_buffer_size / 1024));
         if ((pls->plbuf_buffer = realloc(pls->plbuf_buffer, pls->plbuf_buffer_size)) == NULL)
             plexit("plbuf wr_data:  Plot buffer grow failed");
     }
